Refactor exporters package into logic/commands
