+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -8,7 +8,7 @@ import java.util.Set;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.StringUtil;
-import seedu.address.logic.exporter.FileType;
+import seedu.address.logic.commands.exporter.FileType;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
