+++ src/test/java/seedu/address/testutil/TypicalFileTypes.java
@@ -1,6 +1,6 @@
 package seedu.address.testutil;
 
-import seedu.address.logic.exporter.FileType;
+import seedu.address.logic.commands.exporter.FileType;
 
 /**
  * A utility class containing a list of valid file types to be used in tests.
