+++ src/test/resources/logic/commands/exporter/VcfExporterTest/charlie.vcf
