+++ src/test/resources/logic/commands/exporter/VcfExporterTest/people.vcf
